+++ src/OpCode050B.cpp
@@ -32,7 +32,7 @@ namespace OpenLogReplicator {
             OpCode(oracleEnvironment, redoLogRecord) {
 
         if (redoLogRecord->fieldCnt >= 1) {
-            uint32_t fieldPos = redoLogRecord->fieldPos;
+            uint64_t fieldPos = redoLogRecord->fieldPos;
             uint16_t fieldLength = oracleEnvironment->read16(redoLogRecord->data + redoLogRecord->fieldLengthsDelta + 1 * 2);
             if (fieldLength < 8) {
                 oracleEnvironment->dumpStream << "ERROR: too short field ktub: " << dec << fieldLength << endl;
@@ -49,8 +49,8 @@ namespace OpenLogReplicator {
 
     void OpCode050B::process() {
         OpCode::process();
-        uint32_t fieldPos = redoLogRecord->fieldPos;
-        for (uint32_t i = 1; i <= redoLogRecord->fieldCnt; ++i) {
+        uint64_t fieldPos = redoLogRecord->fieldPos;
+        for (uint64_t i = 1; i <= redoLogRecord->fieldCnt; ++i) {
             uint16_t fieldLength = oracleEnvironment->read16(redoLogRecord->data + redoLogRecord->fieldLengthsDelta + i * 2);
             if (i == 1) {
                 ktub(fieldPos, fieldLength);
