+++ src/OpCode0514.cpp
@@ -39,9 +39,9 @@ namespace OpenLogReplicator {
 
     void OpCode0514::process() {
         OpCode::process();
-        uint32_t fieldPos = redoLogRecord->fieldPos;
+        uint64_t fieldPos = redoLogRecord->fieldPos;
 
-        for (uint32_t i = 1; i <= redoLogRecord->fieldCnt; ++i) {
+        for (uint64_t i = 1; i <= redoLogRecord->fieldCnt; ++i) {
             uint16_t fieldLength = oracleEnvironment->read16(redoLogRecord->data + redoLogRecord->fieldLengthsDelta + i * 2);
 
             if (i == 1) dumpMsgSessionSerial(fieldPos, fieldLength);
